+++ app/src/main/java/rtrk/pnrs1/ra94_2013/projekat1/statisticsNative.java
@@ -1,7 +1,7 @@
 package rtrk.pnrs1.ra94_2013.projekat1;
 
 public class statisticsNative {
-    public native float getStatisticsResult(float mDone, float mTotal);
+    public native float getStatisticsResult(int mDone, int mTotal);
 
     static{
         System.loadLibrary("statisticsNative");
